Provide new requestIsBatched field to gateway compatibility layer (#7118)
This field was added today in v3.11.0 and v4.1.0. That means that if you
installed an older version of Gateway that gets its types from AS3
rather than from `@apollo/server-gateway-interface` it would suddenly
start expecting this field on its GraphQLRequestContext.

We fixed that problem by releasing v3.11.1 which makes the field
optional. But it still seems reasonable to provide it if it's
available... but we keep the field optional in the compatibility
interface so that new Gateway doesn't end up requiring new Server.