Don't override custom code extensions from scalar parseValue errors
Adding `BAD_USER_INPUT` is a nice default (and overrides the
inappropriate default of `INTERNAL_SERVER_ERROR`) but if somebody has
set a `code` already, we shouldn't override.

(Note that there's a separate issue where graphql-js throws out
extensions from the thrown error itself and only pays attention to
extensions on the error's originalError; we're trying to fix that in
https://github.com/graphql/graphql-js/pull/3785 but this is orthogonal.)

Fixes #7178.