Two unrelated small cache improvements (#7241)
- Explicitly allow people to pass `cache: 'bounded'`. Non-TS users
upgrading from the recommended AS3.9+ configuration could do this by
accident. Fixes #7240.

- Upgrade `@apollo/utils.keyvaluecache` so that the new
`PrefixingKeyValueCache.cacheDangerouslyDoesNotNeedPrefixesForIsolation`
feature lets you disable prefixing for the APQ and full response caches.
Throw if you try to pass such a cache to `ApolloServer` itself because
that cache is designed to be shared across features. Migrate off of
PrefixingKeyValueCache for `documentStore` so that its prefixing can't
be disabled. Fixes #6742.