De-comment GitHub templates (#7287)
I believe our intention is that people creating issues/PRs delete this
boilerplate text rather than keeping it in the description. However, if
they don't, it's basically invisible because it's interpreted as a
Markdown comment... but it still ends up in the actual merge commit
message if not explicitly deleted. By de-commenting it, it will be more
obvious if it is left in (and we can just delete it in the UI when we
look at the PR).

This is more relevant for PRs than feature request issues but I think
the point still stands.