Remove unused docs for now (#7293)
These docs were originally kept around with the intention of updating
them for v4 and making them available again in the nav once we did. At
this point, we're not sure when that will happen and we can always
restore them from git history, so I'm deleting for now. This will
hopefully cause navigation to the actual URL to redirect to v3 docs
where they exist, but I'm not sure and might need to follow up with
another PR to add to `_redirects` unless @StephenBarlow knows.