Add arbitrary property to `HeaderMap` for type differentiation (#7314)
Related: https://github.com/apollographql/apollo-server/pull/7313

This shouldn't have been allowed by the typings in the first place.
Previously, a `Map` was compatible with `HeaderMap` according to
TypeScript, so the error wasn't a typing error (when we wanted it to
be). Adding an arbitrary class property to the `HeaderMap` class gives
us the typing error we'd expect when a `Map` is used in lieu of a
`HeaderMap`.