[Docs] AS Integrations AWS Lambda 2.0.0 Update (#7329)
With the v2.0.0 release of `@as-integrations/aws-lambda`, the docs
needed updates to reflect the new syntax and requirements. This PR also
describes more of the added functionality in the 2.0 release.

There was also an issue with the local event type that slipped through
the cracks the first time around due to our runtime type-parsing. The
local event we were using for testing was a `ProxyV1` event type, while
the `httpApi` spun up by `serverless` was of the `ProxyV2` type. I
normalized the example event and integration example to both be the new
`V2` event type (as its now necessary to be explicit with v2.0.0).

Fixes #7327 as `requestContext.http.method` was missing from the event
as it was not required in `ProxyV1`, but is in `ProxyV2`