Show cookie toggle for embedded landing pages (#7358)
[Related embed
PR](https://github.com/apollographql/embeddable-explorer/pull/215)

Context
Users are getting confused because the embedded sandbox doesn't show the
`includeCookies` toggle. This toggle should be visible for AS4 users
when they are in embedded sandbox.

What Changed
This adjusts the way the getEmbeddedHTML helpers use the landing page
config options. These options should stay the same and the user
shouldn't experience any changes to how they need to configure their
Apollo Server configuration.

Instead, we use the `includeCookies` value as an initial value for the
cookie toggle in explorer, pass it to `initialState`, and turn off a
`hideCookieToggle` flag to expose this value to users in explorer.