Add required plugin to allow cache backend to work
I believe that the code in the "Single instance" that calls for redis will not work without the plugin being defined. I've added that along with an appropriate note. Please check me if my assumption is correct. That is, the current code will never work as is.

Also, removed in two places a specific reference to Apollo Server 4 since this should only show up in the Apollo Server 4 doc tree. I assume that when Apollo Server 5 comes out, you do not want to have to go back and change that note (and so on and so forth for every version change)

Note: I assume this change should also be required for the Redis Sential and the Redis Cluster examples, but as this behavior is opaque to me, I really don't know if that is necessary. I would suggest also adding a comment regarding what is really happening and when the responseCachePlugin is needed. I had thoroughly read these doc pages before suggestion this and only know about because of thishttps://github.com/apollographql/apollo-server/issues/7416 discussed with @trevor-scheer.

Another point of confusion for me is that it seems that this section "Cache backends" requires that you have read thoroughly the previous section "Caching" where responseCachePlugin is discussed.  I had read that several times.

It currently says:

You can cache Apollo Server query responses in stores like Redis, Memcached, or Apollo Server's default in-memory cache. For more information, see Configuring cache backends.

And I believe it should say something like

The `responseCachePlugin()` is required to be passed as a plugin along with any KeyvAdapter instantiation. That is, in order to cache Apollo Server query responses in stores like Redis, Memcached, or Apollo Server's default in-memory cache. For more information, see Configuring cache backends.

If this is correct, I'd be happy to update that text, but I feel a little bit like I'm must making stuff up now.