Add Gitleaks allowlist file (#7507)
This PR adds `.gitleaks.toml` to the root of this repo. This file is
consumed by the secret-scanning tool that Apollo's SecOps team uses to
check for secrets in our source.

This file is being added with the required configuration to tell the
scanning tool to ignore values that SecOps has confirmed as false
positives.