Implement callback subscription protocol in new plugin `ApolloServerPluginSubscriptionCallback` (#7617)
This PR creates a new plugin for Apollo Server which implements the new
[subscription callback
protocol](https://www.apollographql.com/docs/router/executing-operations/subscription-callback-protocol/).

While the plugin approach works, it does have a couple notable pitfalls.
I see this as a good first step towards enabling people to use it today.
The nature of this implementation is to skip execution in Apollo Server
altogether, which means the execution internals are bypassed. This means
that some plugin hooks are skipped - notably `executionDidStart/End` and
`willResolveField` (for field-level instrumentation) when resolving
these types of operations.

Fortunately, most of the logic is captured in the `SubscriptionManager`
class. The plugin mostly just wraps it, and Apollo Server should be able
to integrate it when we choose to.