Update list of benign-positive secret detections (#7726)
This PR updates the Apollo-custom Gitleaks allowlist file at the root of
the repo.

Specifically, this PR allowlists 3 detected values which are linked in
comments in the modified TOML file.

A recent update to the version of Gitleaks the Apollo Security team is
using caused additional values to be detected. We have confirmed that
these values are not actually secrets. We would like to ignore these
detections in the future to prevent noise.