feat: allow `stringifyResult` to return a `Promise<string>` (#7803)
* allow `stringifyResult` to return a `Promise<string>`

* call `stringifyResult` in previously missed error response cases as well
Users who implemented the `stringifyResult` hook can now expect error responses to be formatted with the hook as well. Please take care when updating to this version to ensure this is the desired behavior, or implement the desired behavior accordingly in your `stringifyResult` hook. This was considered a non-breaking change as we consider that it was an oversight in the original PR that introduced `stringifyResult` hook.