Add test for router subscription termination via 404 response (#7841)
A user is currently facing an issue where the subscription is not
terminating when the router responds to a `next` request with 404. I
went to verify this behavior in tests, but this case didn't actually
have a test, so I added it.

The interesting branch being tested is corresponds to this log line:
```
"SubscriptionManager[1234-cats]: `next` request received 404, terminating subscription"
```
...which didn't previously exist in any tests before.