Update redis connection docs to match latest Keyv API (#7944)
The `@apollo/utils.keyvadapter` package was updated in
https://github.com/apollographql/apollo-utils/pull/461 to be compatible
with Keyv v5. The primary change is that the `Keyv` constructor takes a
store instead of a string.

The constructor can be invoked in two ways:
```
// passed as first argument
new Keyv(new KeyvRedis("redis://..."))

// passed as store
new Keyv({ store: new KeyvRedis("redis://...") })
```

This updates the Redis examples in the documentation to match the new
API. For simplicity, I've stuck to the first format.