Switch from Volta to Mise for installing dependencies (#8058)
Instead of expecting developers to install a Node-specific dev env tool,
expect a more generic one.

This also lets us use the Mise GH action which automatically caches
tools. (Note that this only runs on main, so it might break in a way we
won't see until we merge.)

This PR slightly rolls back the Node version, primarily so that we can
see that Renovate works.

Also, use a separate CircleCI cache for each Node major version.