Upgrade Nock from v13 to v14
Now that we don't need to support old Node, we can take this upgrade.
This will be helpful because Nock v14 supports mocking the built-in
`fetch` implementation, and we're about to change it so that the
built-in `fetch` is the default fetcher for our plugins.

Two test files that use Nock needed some adjustment:
- `make-fetch-happen` (the default fetcher for `@apollo/gateway`) seems
  to break Nock, so switch the Gateway in the usage reporting test to
  use the built-in `fetch` instead.
- The timing of Nock's `replied` event seems to be slightly different:
  it seems to fire a teeny bit earlier, before the code in Apollo Server
  which processes the result of the mocked-out response. This meant that
  most of the log lines corresponding to AS code receiving responses
  snuck down the snapshot logs a bit, and in some cases meant that we
  stopped our server too fast and caused other issues. Adding a
  `setImmediate` after the `replied` event fixed most of these.