Rewrite standalone server to not use Express
This means that once we drop the middleware from this package, we won't
have a specific version of Express shipping with Apollo Server even if
you don't use Express.

Instead of `express.json` we use `body-parser.json` which is where
Express gets it from.

To generate error pages, we use `finalhandler` directly, which is the
same package Express uses.

Because we only have a fixed composition of four middlewares (cors,
JSON, actual Apollo Server, and error pages) it's straightforward enough
to connect them together directly instead of using Express or
`compose-middleware` or something.

This is an observable change in that the various headers that Express
sets by default like `x-powered-by: Express` and `etag` won't be set. If
you care about this, use Express yourself.

Like most commits on this branch, this doesn't yet update docs.