Initial stab at compatibility with newer GraphQL.JS 17 alphas
AS4 is tested against v16 and v17.0.0-alpha.2.

They've made it to v17.0.0-alpha.8 by now. We don't build against that
alpha for reasons that are fixed by
https://github.com/graphql/graphql-js/pull/4425.

This commit is based on installing a build of that PR (downloaded from
its GH action artifact `npmDist`, unzipped, and `npm pack`'d') and
trying to get tests to pass. These tests should pass with v16,
v17.0.0-alpha.2, and the #4425 custom build.

Incremental delivery tests do NOT pass with the custom build, because
the incremental delivery protocol has changed and we haven't updated
yet.

Smoke test also doesn't pass yet for at least the reason that we haven't
asked it to install the right prerelease.

Changes include:
- Both AS itself and integration tests depend on the particular wording
  of some error messages, which have changed. (Once we can drop v16
  support, AS proper won't need that hack any more, but we're not there
  yet.)
- For some reason we tested the exact size of the JSON serialization of
  a GraphQL-JS AST in documentStore.test.ts.
- One subscriptionCallback debug log line was occurred one tick later
  with the newer module. I added `await setImmediate` so that it now
  *consistently* shows up in that slightly later spot.