Update migration guide for v5
Because we are keeping a single "v4-v5" doc site version, the migration
guide is split into a small "upgrading from v4 to v5" page and a longer
"upgrading from v3 to v5" page based on the previous v3-to-v4 page.

The recommendation in the latter page is to upgrade directly from v3 to
v5; we have no docs specific to going from v3 to v4. (This also means we
no longer have to document the recommendation to set
`status400ForVariableCoercionErrors: true`).

In the latter page, any time we describe a change that occurred between
v3 and v4, we describe it as "in Apollo Server 4+"; however, when
describing how a user will write code today, we describe it as "in
Apollo Server 5".