docs: Add Google Cloud Functions integration (#7929)
Include integration with Google Cloud Functions

This was documented in v3 but disappeared from the list of integrations
in v4. Googling for it was, ironically, an exercise in frustration.

Eventually happened on
https://github.com/apollo-server-integrations/apollo-server-integration-google-cloud-functions

This package is different from the other supported integrations in that
it's not prefixed with `@as/` but instead with
`/apollo-server-integrations` which seems surprisingly similar, but
different. Not sure if this is truly a supported integration or not.

---------

Co-authored-by: Trevor Scheer <trevor.scheer@gmail.com>