Add support for newer incremental delivery format (#8148)
Closes #7932

Adds support for the `graphql@17.0.0-alpha.9` `@defer` and `@stream`
incremental delivery protocol.

When `graphql@17.0.0-alpha.9` is installed, clients must send the
`Accept` header with a value of `multipart/mixed; incrementalSpec=v0.2`
to specify the new format. If the `Accept` header is not compatible with
the installed version of `graphql` (such as sending `deferSpec=20220824`
when ``17.0.0-alpha.9` is installed), an error is returned to the
client.

---------

Co-authored-by: David Glasser <glasser@apollographql.com>