Require legacy incremental execute as an `ApolloServer` option instead of a dynamic import (#8161)
Fixes #8159

Dynamically importing `@yaacovcr/transform` if available has proven to
be a difficult solution because some bundlers fail to build when the
package isn't installed.

This PR addresses this issue by making the legacy execute function an
option that can be provided to the `ApolloServer` constructor. This
means no direct dependency on `@yaacovcr/transform` and makes it
possible to use other legacy execution functions if available.

To migrate, import the `legacyExecuteIncrementally` and provide it as
the `legacyExperimentalExecuteIncrementally` option to the
`ApolloServer` constructor.


```ts
import { legacyExecuteIncrementally } from '@yaacovcr/transform';

const server = new ApolloServer({
  // ...
  legacyExperimentalExecuteIncrementally: legacyExecuteIncrementally
})

```